compose: fix build tags
Compose requires running against a socket.  For native Linux clients
(!remote), we can talk directly to the local Podman socket.  For remote
clients, we need to run against podman-machine which is only available
on amd64 and arm64.

[NO NEW TESTS NEEDED] as there is no functional change.

Fixes: #21757
